feat: refactor header
