+++ src/Header/Header.jsx
@@ -1,0 +1,181 @@
+import React, { useEffect } from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import Menu from '@mui/material/Menu';
+import MenuIcon from '@mui/icons-material/Menu';
+import Container from '@mui/material/Container';
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import Tooltip from '@mui/material/Tooltip';
+import MenuItem from '@mui/material/MenuItem';
+import { Link } from "react-router-dom";
+import { pages } from '../Constants/pages'
+import { useAuth0 } from "@auth0/auth0-react";
+
+// TODO either remove these settings or make pages for them! They should be included in the router (stretch!)
+// const settings = ['Profile', 'Account', 'Dashboard'];
+
+const Header = () => {
+  const [anchorElNav, setAnchorElNav] = React.useState(null);
+  const [anchorElUser, setAnchorElUser] = React.useState(null);
+  const { loginWithRedirect, logout, isAuthenticated, user } = useAuth0();
+
+  const handleOpenNavMenu = (event) => {
+    setAnchorElNav(event.currentTarget);
+  };
+
+  useEffect(() => {
+    console.log(user)
+  }, [user])
+
+  // TODO: lets see about using roles more better - Set user roles in Auth0
+  // https://manage.auth0.com/dashboard/us/<Auth0 Domain>/users
+  const userRole = user && user['http://localhost:3000/role'] && user['http://localhost:3000/role'][0] || 'general'
+
+  const loginLogout = () => {
+    { isAuthenticated 
+      ? <MenuItem
+        onClick={() => logout({ returnTo: window.location.origin })}
+        >
+        Logout
+        </MenuItem>
+      : <MenuItem
+          onClick={() => loginWithRedirect()}
+        >
+        Login
+      </MenuItem>
+    }
+  }
+
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+
+  return (
+    <AppBar position="static">
+      <Container maxWidth={false}>
+        <Toolbar disableGutters>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ mr: 2, display: { xs: 'none', md: 'flex' } }}
+            
+          >
+            HealthyQ
+          </Typography>
+
+          <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
+            <IconButton
+              size="large"
+              aria-label="account of current user"
+              aria-controls="menu-appbar"
+              aria-haspopup="menu"
+              onClick={handleOpenNavMenu}
+            >
+              <MenuIcon />
+            </IconButton>
+            <Menu
+              id="menu-appbar"
+              anchorEl={anchorElNav}
+              anchorOrigin={{
+                vertical: 'bottom',
+                horizontal: 'left',
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: 'top',
+                horizontal: 'left',
+              }}
+              open={Boolean(anchorElNav)}
+              onClose={handleCloseNavMenu}
+              sx={{
+                display: { xs: 'block', md: 'none' },
+              }}
+            >            
+                {pages[userRole].map((page) => (
+                  <MenuItem
+                    onClick={handleCloseNavMenu}
+                    key={page.displayName}
+                    component={Link}
+                    to={`${page.path}`}
+                  >
+                  {page.displayName}
+                  </MenuItem>
+                ))}
+            </Menu>
+          </Box>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}
+          >
+            HealthyQ
+          </Typography>
+          <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+            {pages[userRole].map((page) => (
+              <MenuItem
+                  onClick={handleCloseNavMenu}
+                  key={page.displayName}
+                  component={Link}
+                  to={page.path}
+                >
+                {page.displayName}
+              </MenuItem>
+            ))}
+          </Box>
+
+          <Box sx={{ flexGrow: 0 }}>
+            { isAuthenticated 
+              ? 
+              <>
+                <Tooltip title="Open settings">
+                  <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
+                    <Avatar alt={user.name || 'Unknown'} src={user.picture || "/static/images/avatar/2.jpg"} />
+                  </IconButton>
+                </Tooltip>
+                <Menu
+                  sx={{ mt: '45px' }}
+                  id="menu-appbar"
+                  anchorEl={anchorElUser}
+                  anchorOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                  }}
+                  keepMounted
+                  transformOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                  }}
+                  open={Boolean(anchorElUser)}
+                  onClose={handleCloseUserMenu}
+                >
+                <MenuItem onClick={() => logout({ returnTo: window.location.origin })}>Logout</MenuItem>
+                {/* {settings.map((setting) => (
+                  <MenuItem key={setting} onClick={handleCloseUserMenu}>
+                    <Typography textAlign="center">{setting}</Typography>
+                  </MenuItem>
+                ))} */}
+                </Menu>
+              </>
+            : <MenuItem onClick={() => loginWithRedirect()}>Login</MenuItem>
+          }
+          </Box>
+        </Toolbar>
+      </Container>
+    </AppBar>
+  );
+};
+export default Header;
